Adding mentor service and changes in Mentor DTO, Mentor Entity and Mentor Mapper

+++ microservicio-alumnos/src/main/java/com/mentoresalumnos/model/Mentor.java
@@ -25,7 +25,7 @@ public class Mentor {
     private String apellido;
 
     @Column()
-    private Long edad;
+    private Integer edad;
 
     @Column()
     @ManyToOne
